Crate wasmi_core

Modules

hint

Structs

F32

A NaN preserving f32 type.

F64

A NaN preserving f64 type.

Pages

An amount of linear memory pages.

Trap

Error type which can be returned by Wasm code or by the host environment.

UntypedVal

An untyped value.

Enums

TrapCode

Error type which can be thrown by wasm code or by host environment.

UntypedError

An error that may occur upon encoding or decoding slices of UntypedVal.

ValType

Type of a value.

Traits

DecodeUntypedSlice

Tuple types that allow to decode a slice of UntypedVal.

EncodeUntypedSlice

Tuple types that allow to encode a slice of UntypedVal.

HostError

Trait that allows the host to return custom error.