Struct wasmi_core::F32

pub struct F32(/* private fields */);

A NaN preserving f32 type.

Implementations

impl F32

pub fn from_bits(other: u32) -> Self

Creates a float from its underlying bits.

pub fn to_bits(self) -> u32

Returns the underlying bits of the float.

pub fn from_float(float: f32) -> Self

Creates a float from the respective primitive float type.

pub fn to_float(self) -> f32

Returns the respective primitive float type.

pub fn is_nan(self) -> bool

Returns true if the float is not a number (NaN).

pub fn abs(self) -> Self

Returns the absolute value of the float.

pub fn fract(self) -> Self

Returns the fractional part of the float.

pub fn min(self, other: Self) -> Self

Returns the minimum float between self and other.

pub fn max(self, other: Self) -> Self

Returns the maximum float between self and other.