Struct wasmi_core::UntypedVal

pub struct UntypedVal { /* private fields */ }

An untyped value.

Provides a dense and simple interface to all functional Wasm operations.

Implementations

impl UntypedVal

pub fn to_bits(self) -> u64

Returns the underlying bits of the UntypedVal.

impl UntypedVal

pub fn i32_load( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i32.load Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn f32_load( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the f32.load Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn f64_load( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the f64.load Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i32_load8_s( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i32.load8_s Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i32_load8_u( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i32.load8_u Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i32_load16_s( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i32.load16_s Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i32_load16_u( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i32.load16_u Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load8_s( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load8_s Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load8_u( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load8_u Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load16_s( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load16_s Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load16_u( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load16_u Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load32_s( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load32_s Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i64_load32_u( memory: &[u8], address: Self, offset: u32, ) -> Result<Self, TrapCode>

Executes the i64.load32_u Wasm operation.

Errors

If address + offset overflows.

If address + offset loads out of bounds from memory.

pub fn i32_store( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i32.store Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i64_store( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i64.store Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn f32_store( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the f32.store Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn f64_store( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the f64.store Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i32_store8( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i32.store8 Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i32_store16( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i32.store16 Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i64_store8( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i64.store8 Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i64_store16( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i64.store16 Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i64_store32( memory: &mut [u8], address: Self, offset: u32, value: Self, ) -> Result<(), TrapCode>

Executes the i64.store32 Wasm operation.

Errors

If address + offset overflows.

If address + offset stores out of bounds from memory.

pub fn i32_add(self, rhs: Self) -> Self

Execute i32.add Wasm operation.

pub fn i64_add(self, rhs: Self) -> Self

Execute i64.add Wasm operation.

pub fn i32_sub(self, rhs: Self) -> Self

Execute i32.sub Wasm operation.

pub fn i64_sub(self, rhs: Self) -> Self

Execute i64.sub Wasm operation.

pub fn i32_mul(self, rhs: Self) -> Self

Execute i32.mul Wasm operation.

pub fn i64_mul(self, rhs: Self) -> Self

Execute i64.mul Wasm operation.

pub fn i32_div_s(self, rhs: Self) -> Result<Self, TrapCode>

Execute i32.div_s Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i64_div_s(self, rhs: Self) -> Result<Self, TrapCode>

Execute i64.div_s Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i32_div_u(self, rhs: Self) -> Result<Self, TrapCode>

Execute i32.div_u Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i64_div_u(self, rhs: Self) -> Result<Self, TrapCode>

Execute i64.div_u Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i32_rem_s(self, rhs: Self) -> Result<Self, TrapCode>

Execute i32.rem_s Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i64_rem_s(self, rhs: Self) -> Result<Self, TrapCode>

Execute i64.rem_s Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i32_rem_u(self, rhs: Self) -> Result<Self, TrapCode>

Execute i32.rem_u Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i64_rem_u(self, rhs: Self) -> Result<Self, TrapCode>

Execute i64.rem_u Wasm operation.

Errors

If rhs is equal to zero.

If the operation result overflows.

pub fn i32_and(self, rhs: Self) -> Self

Execute i32.and Wasm operation.

pub fn i64_and(self, rhs: Self) -> Self

Execute i64.and Wasm operation.

pub fn i32_or(self, rhs: Self) -> Self

Execute i32.or Wasm operation.

pub fn i64_or(self, rhs: Self) -> Self

Execute i64.or Wasm operation.

pub fn i32_xor(self, rhs: Self) -> Self

Execute i32.xor Wasm operation.

pub fn i64_xor(self, rhs: Self) -> Self

Execute i64.xor Wasm operation.

pub fn i32_shl(self, rhs: Self) -> Self

Execute i32.shl Wasm operation.

pub fn i64_shl(self, rhs: Self) -> Self

Execute i64.shl Wasm operation.

pub fn i32_shr_s(self, rhs: Self) -> Self

Execute i32.shr_s Wasm operation.

pub fn i64_shr_s(self, rhs: Self) -> Self

Execute i64.shr_s Wasm operation.

pub fn i32_shr_u(self, rhs: Self) -> Self

Execute i32.shr_u Wasm operation.

pub fn i64_shr_u(self, rhs: Self) -> Self

Execute i64.shr_u Wasm operation.

pub fn i32_clz(self) -> Self

Execute i32.clz Wasm operation.

pub fn i64_clz(self) -> Self

Execute i64.clz Wasm operation.

pub fn i32_ctz(self) -> Self

Execute i32.ctz Wasm operation.

pub fn i64_ctz(self) -> Self

Execute i64.ctz Wasm operation.

pub fn i32_popcnt(self) -> Self

Execute i32.popcnt Wasm operation.

pub fn i64_popcnt(self) -> Self

Execute i64.popcnt Wasm operation.

pub fn i32_rotl(self, rhs: Self) -> Self

Execute i32.rotl Wasm operation.

pub fn i64_rotl(self, rhs: Self) -> Self

Execute i64.rotl Wasm operation.

pub fn i32_rotr(self, rhs: Self) -> Self

Execute i32.rotr Wasm operation.

pub fn i64_rotr(self, rhs: Self) -> Self

Execute i64.rotr Wasm operation.

pub fn i32_eqz(self) -> Self

Execute i32.eqz Wasm operation.

pub fn i64_eqz(self) -> Self

Execute i64.eqz Wasm operation.

pub fn i32_eq(self, rhs: Self) -> Self

Execute i32.eq Wasm operation.

pub fn i64_eq(self, rhs: Self) -> Self

Execute i64.eq Wasm operation.

pub fn f32_eq(self, rhs: Self) -> Self

Execute f32.eq Wasm operation.

pub fn f64_eq(self, rhs: Self) -> Self

Execute f64.eq Wasm operation.

pub fn i32_ne(self, rhs: Self) -> Self

Execute i32.ne Wasm operation.

pub fn i64_ne(self, rhs: Self) -> Self

Execute i64.ne Wasm operation.

pub fn f32_ne(self, rhs: Self) -> Self

Execute f32.ne Wasm operation.

pub fn f64_ne(self, rhs: Self) -> Self

Execute f64.ne Wasm operation.

pub fn i32_lt_s(self, rhs: Self) -> Self

Execute i32.lt_s Wasm operation.

pub fn i64_lt_s(self, rhs: Self) -> Self

Execute i64.lt_s Wasm operation.

pub fn i32_lt_u(self, rhs: Self) -> Self

Execute i32.lt_u Wasm operation.

pub fn i64_lt_u(self, rhs: Self) -> Self

Execute i64.lt_u Wasm operation.

pub fn f32_lt(self, rhs: Self) -> Self

Execute f32.lt Wasm operation.

pub fn f64_lt(self, rhs: Self) -> Self

Execute f64.lt Wasm operation.

pub fn i32_le_s(self, rhs: Self) -> Self

Execute i32.le_s Wasm operation.

pub fn i64_le_s(self, rhs: Self) -> Self

Execute i64.le_s Wasm operation.

pub fn i32_le_u(self, rhs: Self) -> Self

Execute i32.le_u Wasm operation.

pub fn i64_le_u(self, rhs: Self) -> Self

Execute i64.le_u Wasm operation.

pub fn f32_le(self, rhs: Self) -> Self

Execute f32.le Wasm operation.

pub fn f64_le(self, rhs: Self) -> Self

Execute f64.le Wasm operation.

pub fn i32_gt_s(self, rhs: Self) -> Self

Execute i32.gt_s Wasm operation.

pub fn i64_gt_s(self, rhs: Self) -> Self

Execute i64.gt_s Wasm operation.

pub fn i32_gt_u(self, rhs: Self) -> Self

Execute i32.gt_u Wasm operation.

pub fn i64_gt_u(self, rhs: Self) -> Self

Execute i64.gt_u Wasm operation.

pub fn f32_gt(self, rhs: Self) -> Self

Execute f32.gt Wasm operation.

pub fn f64_gt(self, rhs: Self) -> Self

Execute f64.gt Wasm operation.

pub fn i32_ge_s(self, rhs: Self) -> Self

Execute i32.ge_s Wasm operation.

pub fn i64_ge_s(self, rhs: Self) -> Self

Execute i64.ge_s Wasm operation.

pub fn i32_ge_u(self, rhs: Self) -> Self

Execute i32.ge_u Wasm operation.

pub fn i64_ge_u(self, rhs: Self) -> Self

Execute i64.ge_u Wasm operation.

pub fn f32_ge(self, rhs: Self) -> Self

Execute f32.ge Wasm operation.

pub fn f64_ge(self, rhs: Self) -> Self

Execute f64.ge Wasm operation.

pub fn f32_abs(self) -> Self

Execute f32.abs Wasm operation.

pub fn f32_neg(self) -> Self

Execute f32.neg Wasm operation.

pub fn f32_ceil(self) -> Self

Execute f32.ceil Wasm operation.

pub fn f32_floor(self) -> Self

Execute f32.floor Wasm operation.

pub fn f32_trunc(self) -> Self

Execute f32.trunc Wasm operation.

pub fn f32_nearest(self) -> Self

Execute f32.nearest Wasm operation.

pub fn f32_sqrt(self) -> Self

Execute f32.sqrt Wasm operation.

pub fn f32_min(self, other: Self) -> Self

Execute f32.min Wasm operation.

pub fn f32_max(self, other: Self) -> Self

Execute f32.max Wasm operation.

pub fn f32_copysign(self, other: Self) -> Self

Execute f32.copysign Wasm operation.

pub fn f64_abs(self) -> Self

Execute f64.abs Wasm operation.

pub fn f64_neg(self) -> Self

Execute f64.neg Wasm operation.

pub fn f64_ceil(self) -> Self

Execute f64.ceil Wasm operation.

pub fn f64_floor(self) -> Self

Execute f64.floor Wasm operation.

pub fn f64_trunc(self) -> Self

Execute f64.trunc Wasm operation.

pub fn f64_nearest(self) -> Self

Execute f64.nearest Wasm operation.

pub fn f64_sqrt(self) -> Self

Execute f64.sqrt Wasm operation.

pub fn f32_add(self, rhs: Self) -> Self

Execute f32.add Wasm operation.

pub fn f64_add(self, rhs: Self) -> Self

Execute f64.add Wasm operation.

pub fn f32_sub(self, rhs: Self) -> Self

Execute f32.sub Wasm operation.

pub fn f64_sub(self, rhs: Self) -> Self

Execute f64.sub Wasm operation.

pub fn f32_mul(self, rhs: Self) -> Self

Execute f32.mul Wasm operation.

pub fn f64_mul(self, rhs: Self) -> Self

Execute f64.mul Wasm operation.

pub fn f32_div(self, rhs: Self) -> Self

Execute f32.div Wasm operation.

pub fn f64_div(self, rhs: Self) -> Self

Execute f64.div Wasm operation.

pub fn f64_min(self, other: Self) -> Self

Execute f64.min Wasm operation.

pub fn f64_max(self, other: Self) -> Self

Execute f64.max Wasm operation.

pub fn f64_copysign(self, other: Self) -> Self

Execute f64.copysign Wasm operation.

pub fn i32_wrap_i64(self) -> Self

Execute i32.wrap_i64 Wasm operation.

pub fn i32_trunc_f32_s(self) -> Result<Self, TrapCode>

Execute i32.trunc_f32_s Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i32_trunc_f32_u(self) -> Result<Self, TrapCode>

Execute i32.trunc_f32_u Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i32_trunc_f64_s(self) -> Result<Self, TrapCode>

Execute i32.trunc_f64_s Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i32_trunc_f64_u(self) -> Result<Self, TrapCode>

Execute i32.trunc_f64_u Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i64_extend_i32_s(self) -> Self

Execute i64.extend_i32_s Wasm operation.

pub fn i64_extend_i32_u(self) -> Self

Execute i64.extend_i32_u Wasm operation.

pub fn i64_trunc_f32_s(self) -> Result<Self, TrapCode>

Execute i64.trunc_f32_s Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i64_trunc_f32_u(self) -> Result<Self, TrapCode>

Execute i64.trunc_f32_u Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i64_trunc_f64_s(self) -> Result<Self, TrapCode>

Execute i64.trunc_f64_s Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn i64_trunc_f64_u(self) -> Result<Self, TrapCode>

Execute i64.trunc_f64_u Wasm operation.

Errors

If self is NaN (not a number).

If self is positive or negative infinity.

If the integer value of self is out of bounds of the target type.

Read more about the failure cases in the WebAssembly specification.

pub fn f32_convert_i32_s(self) -> Self

Execute f32.convert_i32_s Wasm operation.

pub fn f32_convert_i32_u(self) -> Self

Execute f32.convert_i32_u Wasm operation.

pub fn f32_convert_i64_s(self) -> Self

Execute f32.convert_i64_s Wasm operation.

pub fn f32_convert_i64_u(self) -> Self

Execute f32.convert_i64_u Wasm operation.

pub fn f32_demote_f64(self) -> Self

Execute f32.demote_f64 Wasm operation.

pub fn f64_convert_i32_s(self) -> Self

Execute f64.convert_i32_s Wasm operation.

pub fn f64_convert_i32_u(self) -> Self

Execute f64.convert_i32_u Wasm operation.

pub fn f64_convert_i64_s(self) -> Self

Execute f64.convert_i64_s Wasm operation.

pub fn f64_convert_i64_u(self) -> Self

Execute f64.convert_i64_u Wasm operation.

pub fn f64_promote_f32(self) -> Self

Execute f64.promote_f32 Wasm operation.

pub fn i32_extend8_s(self) -> Self

Execute i32.extend8_s Wasm operation.

pub fn i32_extend16_s(self) -> Self

Execute i32.extend16_s Wasm operation.

pub fn i64_extend8_s(self) -> Self

Execute i64.extend8_s Wasm operation.

pub fn i64_extend16_s(self) -> Self

Execute i64.extend16_s Wasm operation.

pub fn i64_extend32_s(self) -> Self

Execute i64.extend32_s Wasm operation.

pub fn i32_trunc_sat_f32_s(self) -> Self

Execute i32.trunc_sat_f32_s Wasm operation.

pub fn i32_trunc_sat_f32_u(self) -> Self

Execute i32.trunc_sat_f32_u Wasm operation.

pub fn i32_trunc_sat_f64_s(self) -> Self

Execute i32.trunc_sat_f64_s Wasm operation.

pub fn i32_trunc_sat_f64_u(self) -> Self

Execute i32.trunc_sat_f64_u Wasm operation.

pub fn i64_trunc_sat_f32_s(self) -> Self

Execute i64.trunc_sat_f32_s Wasm operation.

pub fn i64_trunc_sat_f32_u(self) -> Self

Execute i64.trunc_sat_f32_u Wasm operation.

pub fn i64_trunc_sat_f64_s(self) -> Self

Execute i64.trunc_sat_f64_s Wasm operation.

pub fn i64_trunc_sat_f64_u(self) -> Self

Execute i64.trunc_sat_f64_u Wasm operation.

impl UntypedVal

pub fn decode_slice<T>(slice: &[Self]) -> Result<T, UntypedError>
where T: DecodeUntypedSlice,

Decodes the slice of UntypedVal as a value of type T.

Note

T can either be a single type or a tuple of types depending on the length of the slice.

Errors

If the tuple length of T and the length of slice does not match.

pub fn encode_slice<T>(slice: &mut [Self], input: T) -> Result<(), UntypedError>
where T: EncodeUntypedSlice,

Encodes the slice of UntypedVal from the given value of type T.

Note

T can either be a single type or a tuple of types depending on the length of the slice.

Errors

If the tuple length of T and the length of slice does not match.