Trait wasmi_core::DecodeUntypedSlice

pub trait DecodeUntypedSlice: Sized {
    // Required method
    fn decode_untyped_slice(params: &[UntypedVal]) -> Result<Self, UntypedError>;
}

Tuple types that allow to decode a slice of UntypedVal.

Required Methods

fn decode_untyped_slice(params: &[UntypedVal]) -> Result<Self, UntypedError>

Decodes the slice of UntypedVal as a value of type Self.

Note

Self can either be a single type or a tuple of types depending on the length of the slice.

Errors

If the tuple length of Self and the length of slice does not match.

Object Safety

This trait is not object safe.