Trait wasmi_core::Float

pub trait Float<T>: ArithmeticOps<T> {
    fn abs(self) -> T;
    fn floor(self) -> T;
    fn ceil(self) -> T;
    fn trunc(self) -> T;
    fn round(self) -> T;
    fn nearest(self) -> T;
    fn sqrt(self) -> T;
    fn is_sign_positive(self) -> bool;
    fn is_sign_negative(self) -> bool;
    fn min(self, other: T) -> T;
    fn max(self, other: T) -> T;
    fn copysign(self, other: T) -> T;
}

Float-point value.

Required Methods

fn abs(self) -> T

Get absolute value.

fn floor(self) -> T

Returns the largest integer less than or equal to a number.

fn ceil(self) -> T

Returns the smallest integer greater than or equal to a number.

fn trunc(self) -> T

Returns the integer part of a number.

fn round(self) -> T

Returns the nearest integer to a number. Round half-way cases away from 0.0.

fn nearest(self) -> T

Returns the nearest integer to a number. Ties are round to even number.

fn sqrt(self) -> T

Takes the square root of a number.

fn is_sign_positive(self) -> bool

Returns true if the sign of the number is positive.

fn is_sign_negative(self) -> bool

Returns true if the sign of the number is negative.

fn min(self, other: T) -> T

Returns the minimum of the two numbers.

fn max(self, other: T) -> T

Returns the maximum of the two numbers.

fn copysign(self, other: T) -> T

Sets sign of this value to the sign of other value.