Trait wasmi_core::Integer

pub trait Integer<T>: ArithmeticOps<T> {
    fn leading_zeros(self) -> T;
    fn trailing_zeros(self) -> T;
    fn count_ones(self) -> T;
    fn rotl(self, other: T) -> T;
    fn rotr(self, other: T) -> T;
    fn rem(self, other: T) -> Result<T, TrapCode>;
}

Integer value.

Required Methods

fn leading_zeros(self) -> T

Counts leading zeros in the bitwise representation of the value.

fn trailing_zeros(self) -> T

Counts trailing zeros in the bitwise representation of the value.

fn count_ones(self) -> T

Counts 1-bits in the bitwise representation of the value.

fn rotl(self, other: T) -> T

Get left bit rotation result.

fn rotr(self, other: T) -> T

Get right bit rotation result.

fn rem(self, other: T) -> Result<T, TrapCode>

Get division remainder.